# Optimize distributor tests