<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
Decrease tests execution time

Results in my machine:
Before:
```bash
go test -count=1 ./pkg/distributor/...
ok      github.com/grafana/mimir/pkg/distributor        91.661s
?       github.com/grafana/mimir/pkg/distributor/distributorpb  [no test files]
ok      github.com/grafana/mimir/pkg/distributor/forwarding     0.470s
```
After:
```bash
go test -count=1 ./pkg/distributor/... 
ok      github.com/grafana/mimir/pkg/distributor        34.071s
?       github.com/grafana/mimir/pkg/distributor/distributorpb  [no test files]
ok      github.com/grafana/mimir/pkg/distributor/forwarding     0.171s
```

#### Which issue(s) this PR fixes or relates to

Fixes #2557

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`