@56quarters 2022-07-27
Thanks for this contribution! Instead of spawning goroutines here, can the `t.Parallel()` method be called from inside the test?