@ravilushqa 2022-07-27
> Thanks for this contribution! Instead of spawning goroutines here, can the `t.Parallel()` method be called from inside the test?

Probably yes. But the result will be slower
TestDistributor_PushQuery execution:
before changes: 54sec 
t.Parallel(): 7sec
concurrent: 500ms