@56quarters 2022-07-27
> > Thanks for this contribution! Instead of spawning goroutines here, can the `t.Parallel()` method be called from inside the test?
> 
> Probably yes. But the result will be slower TestDistributor_PushQuery execution: before changes: near 1m
> 

I think that's fine and still an improvement. I'd prefer to use something built into the Go test module instead of spawning manually starting goroutines.