@ravilushqa 2022-07-27
> I think that's fine and still an improvement. I'd prefer to use something built into the Go test module instead of manually starting goroutines.

Make sense, thanks for the review. Changed to `t.Parallel()`